pyproject.toml
[project]
name = "style-sync-backend"
version = "0.1.0"
description = "Style-Sync backend engine (FastAPI + ML image and analytics services)"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
	"fastapi>=0.115.0",
	"httpx>=0.27.0",
	"numpy>=2.1.0",
	"pillow>=10.4.0",
	"python-multipart>=0.0.9",
	"pydantic[email]>=2.0.0",
	"rembg[gpu]>=2.0.59",
	"scikit-learn>=1.5.1",
	"sqlalchemy>=2.0.0",
	"uvicorn[standard]>=0.30.6",
	"python-dotenv>=1.2.2",
]

[project.optional-dependencies]
dev = [
	"pytest>=8.3.0",
	"ruff>=0.6.9",
]

[tool.uv]
package = false
